make Promotion Toggleable, refactor implementation to Trait
