+++ src/Sylius/Component/Promotion/Model/Promotion.php
@@ -13,12 +13,15 @@ namespace Sylius\Component\Promotion\Model;
 
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Collections\Collection;
+use Sylius\Component\Resource\Model\ToggleableTrait;
 
 /**
  * @author Saša Stamenković <umpirsky@gmail.com>
  */
 class Promotion implements PromotionInterface
 {
+    use ToggleableTrait;
+
     /**
      * @var mixed
      */
